# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "afrozsatvilkar2014@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/addbook": {
            "post": {
                "summary": "Add Book",
                "description": "Admin Can Add Book ",
                "operationId": "eda1db9386a16a065c77e2292acdfb26",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "author",
                                    "image",
                                    "Price",
                                    "quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "Price": {
                                        "type": "decimal"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updatebook": {
            "post": {
                "summary": "Update Book",
                "description": "Admin Can Update Book ",
                "operationId": "a04b3bc06167ac6c05240c128d811ab9",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "description",
                                    "author",
                                    "image",
                                    "Price"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "Price": {
                                        "type": "decimal"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book updated Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/addquantity": {
            "post": {
                "summary": "Add Quantity to Existing Book",
                "description": " Add Book Quantity ",
                "operationId": "48b2f793b375e2fb01165ba40c0cce4c",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "quantity"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book Quantity updated Successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deletebook": {
            "post": {
                "summary": "Delete the book from BookStoreApp",
                "description": " Delete Book ",
                "operationId": "831a74bc8ac61c7a30b515d9f82dbf0f",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book deleted Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/displaybooks": {
            "get": {
                "summary": "Display All Books",
                "description": " Display All Books Present in the BookStore ",
                "operationId": "7bf537fdedd60254654c93acd171e750",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Books Available in the Bookstore are"
                    },
                    "404": {
                        "description": "Books are not there"
                    }
                }
            }
        },
        "/api/sortlowtohigh": {
            "get": {
                "summary": "sort on ascending order",
                "description": " sort on ascending order ",
                "operationId": "68ba67ef9f8275d07c2e453e45c83be5",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "These much books are in store ....."
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/sorthightolow": {
            "get": {
                "summary": "sort on Descending order",
                "description": " sort on Descending order ",
                "operationId": "983c1d44dc0db0dbe1e7a75a21e0df03",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "These much books are in store ....."
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/searchbook": {
            "post": {
                "summary": "search the book from BookStoreApp",
                "description": " Search Book ",
                "operationId": "eb87a44bcd1340c50ce83055907bf99d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serch done Successfully"
                    },
                    "403": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}